Fix http issue with optimistic locking.
Using the http driver, when there is no user tx in place, OGM uses http
autocommit mode (just one http request).
This causes problems with optimistic locking, whose semantic is to rollback for
implicit transactions when a concurrent change is detected.

This commit makes the OGM check is the statements use optimistic locking.
If so, then an explicit internal http tx is triggered by ogm, allowing
to rollback in case of error.